Fixed a bug where you couldn't update a Note

See merge request cse1105/2024-2025/teams/csep-team-20!41

+++ server/src/test/java/server/api/NoteControllerTest.java
@@ -177,7 +177,7 @@ class NoteControllerTest {
         var updatedNote = new Note("Title2", "2");
 
 
-        updatedNote = noteController.updateNote(1,updatedNote).getBody();
+        updatedNote = noteController.updateNote((long) 1,updatedNote).getBody();
 
         assertEquals(updatedNote, noteController.getById(1).getBody());
     }
